backend-adk/requirements.txt
# Copyright 2025 Google LLC.
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#     https://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

absl-py==2.1.0
google-adk==1.5.0
requests==2.32.4
uvicorn==0.34.0
google-auth==2.37.0
fastapi==0.115.14
google-api-core==2.24.0
google-cloud-logging==3.11.3
pydantic==2.11.4
google-generativeai==0.8.5
tadau==1.0.8
bs4==0.0.2
google-streetview==1.2.9
Deprecated==1.2.18
aiohttp==3.12.13
python-dotenv==1.1.0
googlemaps==4.10.0
mimetype==0.1.5
google-cloud-storage==2.19.0
pytest==8.4.1
pytest-asyncio==1.0.0
pytest-cov==6.2.1